+++ CHANGELOG.rst
@@ -12,6 +12,7 @@ Added
 - On the “Entries” tab, the data views are expanded to include “Results”.
 - On the “Entries” tab, the current ranking is displayed in all data views.
 - On the "Entries" tab, the option to print results and split times has been added.
+- Display of competitor data grouped by 'Competitors' and 'Clubs' added.
 
 
 [0.4.8] - 2026-04-26